Create README.md
The website helps the farmers with veterinarian services
without the need of visiting the actual veterinary care in
rural areas. The users can contact veterinarians, get
feedback on various diseases, know about animal
healthcare and hygiene, and can also schedule
appointments in veterinary care. For this, the user just
needs to access the website through a proper internet
connection and therefore the system provides costeffective
services which can be accessed easily at any
time and any place. If such a system is used in Gram
Panchayats of villages it can reach each farmer and help
them to provide proper treatment to their farm animals.
A good amount of information is also added to the
website divided into Disease portals and Health
Lifestyle Management portals for the Livestock on the
website. Thus, successfully implemented animal
healthcare facilities and disease prediction systems as a
website. Through the Disease prediction portal farm
animal owners can predict disease more precisely which
would help in the treatment of the animal. Using ML an
effective Disease prediction system can be made that
will help livestock owners with the prediction of
different diseases found in animals. Prediction of
diseases will help provide proper treatment on time.
Accurate predictions by such a model will provide aid
to maintaining a healthy animal lifestyle which in turn
will boost the production of livestock products that will
satisfy the increasing demands.